DLP policy rule
In the Message trace details the Transport Rule event could be for a DLP policy OR Transport Rule. For example the details will display the following - 

Apply HTML disclaimer. Transport rule: ",ID: ('B58239483-38792-7837-8738-324389'), DLP Policy ", ID: (000000-0000-0000-0000-0000000).

In this case a Transport rule hit the message to apply a disclaimer and no DLP policy was hit. This is different than the custom-data field which will only show the Transport rule or DLP policy details, not blank information.

Thanks,
Kelly